Fix if enabled switches
There was an copy and paste error in the include logic if ldap.enabled
was set to true. So everytime the ldap block got included. This went
undetected as we use LDAP for authentication.

Another switch was added when we discovered that there is an special
-ldap tagged version of the netbox docker image. Now the deployment will
fetch the special tagged ldap enabled image, when setting
.Values.ldap.enabled to true

Signed-off-by: Marcus Franke <marcus.franke@gmail.com>