feat(netbox): allow runtime configuration
adds the option to append to the runtime configuration.py script from
the values file. this is helpful for some plugins (namely netbox
branching) which require some runtime changes. additionally, plugins
like netbox diode require us to read in some new information that are
secrets (namely, the diode to netbox client token), but cannot be
mounted _and_ configured through helm. an example of the additional
config is given below:

```
configBootstrap: |
  # Bootstrap netbox branching plugin
  NETBOX_BRANCHING_PLUGIN = "netbox_branching"
  if NETBOX_BRANCHING_PLUGIN in PLUGINS:
      from netbox_branching.utilities import DynamicSchemaDict

      DATABASES = DynamicSchemaDict(DATABASES)
      DATABASE_ROUTERS = [
          'netbox_branching.database.BranchAwareRouter',
      ]
```